# openapi.yaml
openapi: 3.1.0
info:
  title: NaijaMove API
  version: 1.0.0
  description: |
    REST API for the NaijaMove ride-hailing platform.
    All monetary amounts are in **kobo** (1 NGN = 100 kobo).
    All timestamps are ISO 8601 UTC strings.
  contact:
    name: NaijaMove Engineering
    email: engineering@naijamove.com
servers:
  - url: http://localhost:3004
    description: Local development
  - url: https://api.naijamove.com
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /drivers/register:
    post:
      tags: [Drivers]
      summary: Register as a driver
      description: |
        Called by a signed-in **rider**. Creates a `pending` driver profile, switches the
        account role to `driver`, and returns a fresh token pair carrying that role —
        replace the stored tokens with these. Admin approval is still required before
        the driver can go online.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Driver profile created and role upgraded
          content:
            application/json:
              example:
                driver:
                  id: "drv_01HZ9K2M3N4P5Q6R7S8T9U0V"
                  userId: "usr_01HZ9K2M3N4P5Q6R7S8T9U0V"
                  status: "pending"
                  createdAt: "2024-01-15T10:30:00Z"
                accessToken: "eyJhbGciOiJIUzI1NiIs..."
                refreshToken: "eyJhbGciOiJIUzI1NiIs..."
        "403": { description: Caller is not a rider }
        "409": { description: Driver profile already exists }
  /drivers/me:
    get:
      tags: [Drivers]
      summary: Get my driver profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Driver profile
          content:
            application/json:
              example:
                id: "drv_01HZ9K2M3N4P5Q6R7S8T9U0V"
                name: "Emeka Okafor"
                phone: "+2348025550193"
                rating: 4.87
                totalTrips: 1240
                status: "approved"
                isOnline: true
                tier: "gold"
                acceptanceRate: 94
                completionRate: 98
  /drivers/availability:
    post:
      tags: [Drivers]
      summary: Set online / offline status
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isOnline]
              properties:
                isOnline:
                  type: boolean
            example:
              isOnline: true
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              example:
                isOnline: true
                updatedAt: "2026-09-11T16:00:00Z"
  /drivers/location:
    post:
      tags: [Drivers]
      summary: Update driver GPS location
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat:
                  type: number
                  minimum: -90
                  maximum: 90
                lng:
                  type: number
                  minimum: -180
                  maximum: 180
            example:
              lat: 6.4281
              lng: 3.4219
      responses:
        "200":
          description: Location updated
  /drivers/earnings:
    get:
      tags: [Drivers]
      summary: Get driver earnings summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Earnings breakdown
          content:
            application/json:
              example:
                today:
                  tripsCompleted: 7
                  grossKobo: 620000
                  commissionKobo: 124000
                  netKobo: 496000
                week:
                  tripsCompleted: 34
                  grossKobo: 3460000
                  commissionKobo: 692000
                  bonusKobo: 200000
                  netKobo: 2968000
                walletBalanceKobo: 1260000
  /drivers/documents:
    get:
      tags: [Drivers]
      summary: List my documents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Document list
          content:
            application/json:
              example:
                documents:
                  - id: "doc_01HZ9K2M3N4P5Q6R7S8T9U0V"
                    type: "drivers_licence"
                    status: "verified"
                    expiresAt: "2028-06-02T00:00:00Z"
                  - id: "doc_02HZ9K2M3N4P5Q6R7S8T9U0V"
                    type: "insurance"
                    status: "expiring_soon"
                    expiresAt: "2026-09-25T00:00:00Z"
    post:
      tags: [Drivers]
      summary: Record a document
      description: |
        Upload the file first with `POST /uploads/presign` (purpose `driver_document`)
        and PUT the bytes to the returned URL; then submit the `key` here as `fileKey`.
        The server verifies the key was issued to you for this purpose and that the
        object exists before storing it.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [drivers_licence, vehicle_registration, insurance, inspection, background_check, profile_photo]
                fileKey:
                  type: string
                  description: Object key from /uploads/presign
                referenceNumber:
                  type: string
                expiresAt:
                  type: string
                  format: date-time
            example:
              type: "drivers_licence"
              fileKey: "driver-documents/usr_01HZ9K2M3N4P5Q6R7S8T9U0V/3f1c…/a1b2.jpg"
              referenceNumber: "LIC-2024-00482"
              expiresAt: "2028-06-02T00:00:00Z"
      responses:
        "200":
          description: Document recorded (status pending review)
        "403":
          description: fileKey was not issued to this user for driver documents
        "422":
          description: File has not been uploaded yet
  /drivers/documents/{documentId}/url:
    get:
      tags: [Drivers]
      summary: Short-lived read URL for a document
      description: Owner (driver) or admin. URL expires in 10 minutes.
      security:
        - bearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Presigned GET URL
          content:
            application/json:
              example:
                url: "https://s3.us-west-004.backblazeb2.com/naijamove-uploads/driver-documents/…?X-Amz-Signature=…"
                expiresAt: "2026-09-13T12:10:00Z"
        "403": { description: Not your document }
        "404": { description: Document not found }
  /drivers/me/photo:
    post:
      tags: [Drivers]
      summary: Attach an uploaded profile photo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileKey]
              properties:
                fileKey: { type: string, description: Object key from /uploads/presign (purpose profile_photo) }
      responses:
        "200":
          description: Photo attached
          content:
            application/json:
              example: { avatarKey: "profile-photos/usr_…/….jpg" }
  /drivers/{driverId}/documents:
    get:
      tags: [Drivers]
      summary: List a driver's documents (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: driverId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Documents for review
        "404":
          description: Driver not found
...
  /drivers/{driverId}/approve:
    post:
      tags: [Drivers]
      summary: Approve a driver (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
          example: "drv_01HZ9K2M3N4P5Q6R7S8T9U0V"
      responses:
        "200":
          description: Driver approved
          content:
            application/json:
              example:
                id: "drv_01HZ9K2M3N4P5Q6R7S8T9U0V"
                status: "approved"
  /drivers/{driverId}/suspend:
    post:
      tags: [Drivers]
      summary: Suspend a driver (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
            example:
              reason: "Multiple rider complaints about unsafe driving"
      responses:
        "200":
          description: Driver suspended

# ─────────────────────────────────────────────────────────────────────────────
# VEHICLES
# ─────────────────────────────────────────────────────────────────────────────
...
  /rides/driver/trips/{tripId}/arrive:
    post:
      tags: [Rides]
      summary: Mark arrived at pickup (driver)
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Arrival recorded
  /rides/driver/trips/{tripId}/start:
    post:
      tags: [Rides]
      summary: Verify PIN and start trip (driver)
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties:
                pin:
                  type: string
                  minLength: 4
                  maxLength: 4
            example:
              pin: "4821"
      responses:
        "200":
          description: Trip started
  /rides/driver/trips/{tripId}/complete:
    post:
      tags: [Rides]
      summary: Complete a trip (driver)
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            example:
              finalDistanceMeters: 8650
              finalDurationSeconds: 1620
      responses:
        "200":
          description: Trip completed and earnings settled
          content:
            application/json:
              example:
                status: "completed"
                grossKobo: 185000
                commissionKobo: 37000
                netKobo: 148000
...
  /payments/payout/driver:
    post:
      tags: [Payments]
      summary: Trigger driver payout (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              driverId: "drv_01HZ9K2M3N4P5Q6R7S8T9U0V"
              amountKobo: 500000
              settlementCycleId: "cyc_01HZ9K2M3N4P5Q6R7S8T9U0V"
      responses:
        "200":
          description: Payout created
...
  /maps/route:
    post:
      tags: [Maps]
      summary: Get route distance and duration
      description: OSRM-based routing (OSM). Call this before /pricing/quote.
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            example:
              pickupLat: 6.4281
              pickupLng: 3.4219
              destinationLat: 6.6018
              destinationLng: 3.3515
      responses:
        "200":
          description: Route result
          content:
            application/json:
              example:
                distanceMeters: 8400
                durationSeconds: 1560
                polyline: "_p~iF~ps|U_ulLnnqC_mqNvxq`@"

# ─────────────────────────────────────────────────────────────────────────────
# SAFETY
# ─────────────────────────────────────────────────────────────────────────────
...
  /admin/drivers:
    get:
      tags: [Admin]
      summary: List all drivers (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending_review, approved, suspended, rejected]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Driver list
  /admin/drivers/{driverId}/approve:
    post:
      tags: [Admin]
      summary: Approve a driver (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Driver approved
  /admin/drivers/{driverId}/suspend:
    post:
      tags: [Admin]
      summary: Suspend a driver (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            example:
              reason: "Multiple safety complaints"
      responses:
        "200":
          description: Driver suspended
...